feat: minor refactor
